tapfreighter: move funding related code into its own file
This is a pure code move. We're going to add unit tests for the moved
functions in the next commit.
